Improve error handling.
We maintain a stack of errors in the global state. Each new error
is pushed onto this stack and marked as active. When a UI element
handles an error, it marks the error as handled through an action.
A result of this design is that all errors that have ever occured
are retained in state (but marked as inactive). We likely want to
add a bound to this stack so that accumulated errors don't slow
down the app.