Add OAuth functionality (currently iOS only). (#585)
This diff enables OAuth based authentication with the Zulip server. A one time pad is generated and sent to the server. The server processes the OAuth request inside of an embedded browser view and returns control to the app via a custom URI scheme (zulip://). The response contains the authenticated API key XORed with the one time pad in order to prevent URI hijacking.