Fix AutoComplete from showing up when typing Emails
The AutoComplete View pops up when typing '@' to mention usernames, but when typing email addresses the same View pops up. This was fixed by checking whether the previous letter before the '@' is an empty space. This stops the view from showing as the letter before '@' would never be a space in an email.

Fixes:#1553