docs: Refactor "getting started" material.
We had a "Getting Started" doc that was mostly about commands to use
day to day and didn't say how to get started, and a "Developer Guide"
with lengthy details about getting started on each of Android and iOS.

Reorganize all the material into a hopefully more helpful arrangement:
a single entry point, with links to one-time setup instructions, the
latter mostly in separate per-platform docs.

There are still some missing steps here, some I know of and I suspect
some others; leave that for separate commits.  This change is a pure
refactor that just moves the material around.