docs: Fix instructions on dev dependencies.
Yarn expects Node to be installed already, and it took me a bit of
riddle-solving to determine the right way to do that and the right
Node version to install.  So, document that.

Also, move the line about making a server dev VM down to the section
that's about making that work all the way -- and clearly label that as
optional.