docs: Dedupe Android and iOS setup instructions.
Now that we're mostly just pointing to React Native upstream's
instructions, we can just do that once.  Then the Android-specific
tips are short enough to inline back into the main doc.

Also this fixes the lie we were telling where the dev setup
"should be as simple" as a `git clone` and `yarn install`, or
the softer version saying "just".  As the rest of these docs
make clear, the setup is not that simple. :-p