mainNavBar: Fix re-render on navigating back.
Use connectWithActionsPreserveOnBack instead
of connectWithActions so that it do not
re-render unnecessary on navigating back.

This can be tested easily on Android, little bit old device.

It fixes the momentary flash of white color (from stream color)
in nav bar on navigating back to main screen.

In private narrow, back arrow is not visible for some millisecond
in transaction back to main screen.