dev auth: Fix crash by using `!!` in a JSX short-circuit expression.
The app was crashing immediately on entering this screen.  Examining
the device's log using `adb logcat` produced a traceback from deep
inside React Native, with this cryptic error message:

  java.lang.RuntimeException: Cannot add a child that doesn't have a
  YogaNode to a parent without a measure function! (Trying to add a
  'ReactRawTextShadowNode' to a 'LayoutShadowNode')

From https://github.com/facebook/react-native/issues/13243 , we learn
a crash like this can result from a variety of errors in using JSX.

Here, the issue seems to be that when `error` is the empty string,
we would stick that empty string directly in as a child of the View.
Applying the classic bang-bang operator to convert to a boolean
fixes the issue.