screen: Add a optional prop keyboardShouldPersistTaps for scrollView.
By default, when a text input inside a ScrollView is focused and the
user taps outside it, the keyboard is dismissed and the tap doesn't
reach any children of the ScrollView.  On certain screens in the
onboarding and login flow, we want to disable this behavior so that
the user can tap directly on a button to submit the input.  Add a prop
to control this, which we pass directly to the ScrollView.

See upstream docs for this prop:
https://facebook.github.io/react-native/docs/scrollview.html#keyboardshouldpersisttaps

[greg: Revised commit message.]