screen: Add a optional prop keyboardShouldPersistTaps for scrollView.
This prop is used to determines when the keyboard should stay visible after a tap.

It can be one of from never, always, handled.

* 'never' (the default), tapping outside of the focused text input
when the keyboard is up dismisses the keyboard.
When this happens, children won't receive the tap.

* 'always', the keyboard will not dismiss automatically,
and the scroll view will not catch taps,
but children of the scroll view can catch taps.

* 'handled', the keyboard will not dismiss automatically
when the tap was handled by a children, (or captured by an ancestor).

https://facebook.github.io/react-native/docs/scrollview.html#keyboardshouldpersisttaps