dev: Make hot reload work again by bumping a dependency.
Hot reload is a pretty crucial feature for iterating on the app at a
good pace, and it wasn't working (at least for me, in my Android
emulator).  So I spent some time digging into it for a solution.
This fix comes from a comment here, and it works for me:
  https://github.com/facebook/react-native/issues/16505#issuecomment-367482457

Some later comments in that thread report the same fix not working;
it seems like React Native's implementation of this feature is pretty
fragile, unfortunately.  Hopefully this fixes it in our project for
all contributors, though, at least until the next time it breaks.