Added initial build directories (generated/,intermediates/ and outputs/) to .gitignore.
While generating initial build(generating the android apk) for mobile app and setting up the project. Several build and intermediate files get generated automatically that are not setup to be ignored by git. As a result intermediate files end up appearing as changed files in working directory and it becomes difficult to keep track of changes with Git. This commit aims to avoid this problem by ignoring those newly generated files.