emoji: Include Zulip extra emoji in the project.
Currently three types of emojies are supported in
reactions and messages:
https://github.com/zulip/zulip/blob/025956482a62cd7f5e2a344bb258b953c7699b8c/zerver/models.py#L1293
* unicode
* realm emoji
* zulip extra emoji

Unicode & realm emoji are handled nicely, but zulip extra emoji
was completely ignored. So in this commit I have added emojies
to the project and in futher commits will be using it to render
them in message list and autocomplete pop up.