WIP: Tried experimental workarounds to load images into chat webview
The aim of this commit is only to enlist possible attempts and failures occured when I tried to load images from backend.
Various workarounds were tried and all the present within this commit.
Various sections of the code has been commented out many of which contain console.logs() calls which are log commands for reactotron and are placed at specific locations to get information about various sections of code.
One more thing to note is, most of the code is static and is just written to illustrate the least and simplest working case in order to test if something truely works. Much of the code might require a little editing for user specifig senario kindly keep that in mind.
Further note that I am not sure if my implementations are correct, so If I have tried something in a way that it is not meant to be done, kindly correct me.