webvew: Batch webview input messages
Send input messages as array of objects instead of several times
as a single object.

That increases performance, reliability and predictability.

There are several reports in the GitHub issues for React Native
that sometimes messages are not delivered when several sent in
quick succession. I could reproduce that, though very rarely.
Hopefully this code fixes that issue, but even if not, the
benefits remain.