baseSelector: Return data with undefined value in default nav params.
Like narrow is undefined in default chat screen params, it's better
to do same for other screen params. So that if params value is undefined
then we can conclude that it need to be created instead of creation.
Like create stream and edit stream screen.

Fix: #2288 stream creation is broken.