streamLink: Fix navigation to streams from messages issue.
Streams have been started to named with `streamId`
appended to the front.

New url format:
/#narrow/stream/{stream_id}-{stream_name}.

To properly navigate to streams from in-message # links,
(e.g. #mobile), we need to extract the stream_name.

Fixes #2312

streamLink: Test extractStreamID with multiple variations of input.

The method `extractStreamID` does not take into account
the "validness" of the extracted number.
(e.g. that number may not be present in the streams array
as a valid `stream_id`. )

This is an intended behavior to abstract the high level logic
into smaller methods. Hence, "validness" is not tested.