reactions: Fix clicking on reactions doesn't work.
This was due to migrating to latest add/remove
reations api (r'^messages/(?P<message_id>[0-9]+)/reactions$')
from (r'^messages/(?P<message_id>[0-9]+)/emoji_reactions/(?P<emoji_name>.*).

Migrations was done in #1896.

Former was only dependent on emoji name, but latest one
requires emoji code and emoji type as well.

Now when app is open and and emoji add/remove event is
arrived, it doesn't contain emoji type and code in it's
payload. So we weren't able to send back code & type
on user clicking on reactions.

But when a fresh message is fetched from the server,
it contains all information, so that it was working.
And was not 100% reproducible.

In this commit, some utils method are created which helps
in finding the emoji code & type.

Now clicking on reactions works as expected.

Fix: #2116