webview: Add API key to images' URLs so they can load.
The Zulip server generally requires authentication to view uploaded
images.  This has long been true of the S3 uploads backend, which we
use on zulipchat.com, and with 1.8 it's now also true of the local
backend (except on Ubuntu Trusty, e.g. on chat.zulip.org.)  So we need
to provide authentication.  Until now, we've never had this working in
the WebView-based message list.

React Native unfortunately doesn't support HTTP basic authentication
inside a WebView, even though both the underlying Android and iOS SDKs
do.  So we can't use that like we do for our API requests.

We could acquire a session cookie, which is how the webapp
authenticates its requests to the server.  But the API keys the app
already uses don't apply quite as broadly as our session cookies
(e.g., they can't be used to deactivate the user's account), so doing
that would require either carefully reviewing the different security
scopes or creating a second class of session cookies, either of which
would take quite a bit of work.

Instead, we've added to the server an alternate form of authentication
which is accepted only for the `/user_uploads/` route: the API key can
be passed as a query parameter.  See
  https://github.com/zulip/zulip/commit/5ddf2614f

Use that here.

[greg: fixed an overrun bug and a non-escaping bug in the regex,
 adding tests; expanded commit message.]