url: Replace is-url with url-regex for precise url validation
is-url validates only 'loosely' according to their own description.
Some important cases were missed (double dots etc.) and on
Android this results in a crash.

url-regex uses a very well tested and almost 'famous' regex:
https://gist.github.com/dperini/729294

As seen here it tests all tests but the '_' one:
https://mathiasbynens.be/demo/url-regex

We do not allow underscores in our subdomains and it is
very uncommon character to be used, also not valid in domain names.