streamLink: Add support for both legacy and new stream names.
To properly navigate to streams (also topics and conversations)
from in-message # links, (e.g. #mobile, #general),
we need to first check presence of stream_id in the url.

If indeed it is present and valid,
we can do quick lookup the latest `stream_name`
associated with that `stream_id`.
When we have the latest `stream_name`,
we can continue to do operations with it.

With this solution, we don't have to change the type of `operand`
or structure of the `doNarrow()`.