search message screen: Fix messageList is not visible.
Actually the issue is: webView is not visible if it's direct
or indirect parent is scrollView. It can be easily tested
by surrounding message list with scrollView in chat.js.

We can make it visible by applying flex:1 style to scrollView
contentContainerStyle, but then it will broke scroll.

So in this commit this style is only applied for search message
screen.