refactor: Do not store new messages in narrows not open
For every narrow we have data for that is caughtUp at 'newer'
we store new incoming messages from the event polling.

Previously we determined if a message should be put inside a
narrow even if no messages were loaded for it. This is not needed
and even does not serve as a good cache because on narrowing
we throw it away and reload new messages.

This also simplifies the code and makes it more predictive.