webview: Cut buggy and now-obsolete realm-prefix hack.
This hack is a bit overzealous; if another user sent a message with
a certain pattern of ordinary text in it, like the message

  src="/

we would go ahead and insert the realm URL into what our user saw.

Rather than fix the hack, just cut it entirely -- now that we pass
`baseUrl` to the WebView, there's no need to try to absolutize
relative URLs.