keyboard: Do not 'swallow' first taps in users screen.
We open Users screen with a keyboard open. That is OK.
But then tapping on a user as a first action hides the keyboard
but does not narrow to the user.

Instead, tapping the user now narrow to that user, and tapping outside the
user list will hide the keyboard.