Fix isFetching selector logic
'needsInitialFetch' is a flag used to determine if we should fetch
data (via fetchEssentialInitialData and fetchRestOfInitialData)

The logic of 'isFetching' might have been correct in the past but now we
are using it differently. Remove the 'needsInitialFetch' flag from the
selector and keep only the fetching older & newer for current narrow.