ui: Change default display for UnreadCunt
We are 'limiting' the display count in the UnreadCount component
in a way so if the value is more than 99 it just shows '99+'.

This is still potentially useful so keep the behavior as an option,
but by default do not do that and instead show the full count.

This is to make the behavior consistent with the web app.