Simplify AccountDetails screen
This was a not so successful experiment at providing different views
for different screen orientations. While this works, the complication
and extra code adds too much to our maintenance burden versus the
benefit this code adds (I several times did not change stuff in this
code because it was twice the work)

As a first step, just merging the portrait code into AccountDetails and
completely scraping the 'landscape' one.