refactor: Remove mapApiToStateUser
We started using this concept in the beginning, matching the
API response to an internal more JavaScript-friendly format, that
was optionally could contain more data and be less API-dependent.

This turned out to be too much work, for too little benefit,
especially with the use of selectors.

The use of this remained for the users data structure and for
nothing else (we had User and ApiUser types)

* merge User and ApiUser into a single type
* do not map from ApiUser to User but keep data as-is
* replace all use of camelCase to snake_case