webview: Escape HTML in displayed errors.
In addition to the code in src/webview/html/, there's this one bit in
js.js where we build up some HTML.  Escape the data we interpolate
here too.

Because this code runs inside the webview and can't import from
our main codebase, and there's only a little of it, skip the fancy
template tag in favor of manually invoking an escape function.
On the other hand, take advantage of the browser environment to
implement that escape function very simply.