Fix jerkiness when opening/closing compose menu
This was caused by several lines of text being formed while the compose input box narrowed to accomodate the menu.
Now, on tapping the menu expand button, the height of input box is immediately reduced to a fixed value.
When menu is closed, the height remains fixed till the menu closing animation is complete, and then is replaced by height: auto.

Two new styles have been defined in composeBoxyStyles.js for this purpose: composeTextInputShowFullText and composeTextInputShowOneLine.

Fixes #2572