refactor: Use easier `import` for actions in ComposeBox
We consolidate and export all actions in actions.js
This allows for shorter import of actions, that we need to do with
calling `dispatch`.

There is an even shorter option as used in `connectWithActions`:

```js
import * as actions from './actions';
```

Then we use it as:
```js
dispatch(action.addToOutbox(...));
```

I am not sure if this is an improvemet or not, but is an option.