flow: Add concrete type for EventUpdateMessageAction
This is the event fired when a message on the server have been updated.

There are three cases that could cause that:
* message's content have changed
* message's topic have changed
* both content and topic have changed

Curenly we are handling only the content change and partially the
topic change.

We are also not doing anything with the `propagate_mode`.

These types are correct but the tests are failing because the reducer
was not aware of all the incoming data formats (it either changed
since we implemented it, or we implemented it incompletely).

Now thanks to Flow we know what remains to be done.