Add `updateTextInput` for direct TextInput text change
Instead of using the topic and message inputs' property `value`
we can short-circuit this and update native properties ourselves.

There are certain very rare cases, because of the async-nesss of
the communication between React Native and the underlying native
layer that might result in a call on an invalid reference.

Thus we add a check before the call to `setNativeProps`.