common/input: Adjust underline color on Android.
The underline's color had been a darker green when focused, and black
when blurred.  BORDER_COLOR is more appropriate, and HALF_COLOR
properly reduces emphasis in light mode and provides visibility in
dark mode.

This has no effect when the Input's props (passed through as
`restProps`) override `underlineColorAndroid`, as when we set it to
"transparent" (meaning no underline) for the compose and topic
inputs and a few others.