store: Include `users` in the persisted state.
This fixes #2591, where tapping on a notification for a group PM
sometimes led to the 'All messages' narrow, instead of the correct
one.  The cause of that issue was that without the `users` state, we
didn't have the mapping of those user IDs to their email addresses,
and much of the app still relies on email addresses to identify users.

Alternatively we could have waited asynchronously for the `users` data
to load from the server, but that would require a lot of rework and
lengthen the action to potentially several seconds on a mobile network.

Initially we were preserving most of the store keys, but we've cut out
many of them to reduce the time spent rehydrating at startup.  We cut
out `users` in commit ecced5bca last October.

[greg: rewrote commit message]