Integrate 'clearMessageInput' functionality in 'updateTextInput'
React Native has problems with some custom Android keyboards. It
fails to reset the TextInput contents. To fix that we use
`react-native-text-input-reset`

In the move to non-controlled inputs in ComposeBox it makes sense
to integrate this reset with the 'updateTextInput' functionality,
both for simplicity and to reuse for the topic input.