dev auth: Replace broken dev_get_emails endpoint with dev_list_users.
The `dev_get_emails` Zulip server endpoint was renamed to
`dev_list_users` in zulip/zulip:383c62fb0.  The sole purpose of this
endpoint is to supply the mobile app, in the dev environment, with a
list of users for the dev login screen.  Because it got renamed, the
mobile dev login view broke.

This commit uses the new endpoint and also addresses changes
in the data structure returned by the endpoint.

The extra data in the new endpoint is intended to allow our dev login
flow to let us log in as users in various realms, rather than
defaulting to a particular one.  This change doesn't yet update the
flow to actually do that; just to continue working, with the
functionality it had before the server-side change.