autocomplete: Factor `**` markers out from `getAutocompletedText`.
Previously `getAutocompletedText` was filling in the `**`s around
the autocompleted users and streams. It is unable to differentiate
between a user and a user group, so make it simpler and do not
add these special symbols, but expect them to be added before
calling the function.

This lets us also fix its newest call site, for user groups, to fill
in a single `*` instead, which is the correct marker for those.

[greg: Squashed change to function with change to call sites,
 and merged commit messages.]