ui: Fix AccountItem delete button alignment
Make sure that the delete button is aligned vertically.
Removing the `padding` style works, but we do this instead:

```js
padding: HALF_SPACING,
margin: -HALF_SPACING,
```

The default button size is too small - 32px (with the minimal
suggested width being 44px, and the larger the better)

Padding increases the touch area (now 32 + 8 + 8 = 48, good enough)
and a negative margin keeps the button at the same place as if it
did not have padding set.