docs/debugging: Be specific about Chrome dev tools.
The upstream RN doc we link to is pretty specific in referring to the
Chrome developer tools, not a generic browser's.  I'm not sure how
much it's really specific to Chrome, or to what extent it would also
work in e.g. Firefox; but let's follow upstream's doc unless someone
specifically explores that in more detail.