account switcher: Adjust layout toward Material guidelines.
The trash icon is annoyingly big, with too small a touch target.
Also, on iOS though not on Android, it renders below the
vertically-centered spot where it should be.  Fix those issues.

For reference for specific measurements, we look to Material Design.
This component is a list item, with two lines of text and a right-hand
icon, so refer to the section "Double line" here:
  https://material.io/design/components/lists.html#specs

Other changes include:
* The text lines get 4px of padding between them.
* The items get a bit taller, and the corner radius adjusted to 4px.