account pick screen: Use standard Redux dispatch.
See #2628 for context.

A thing specific to this commit is that the `dispatch` call
inside the `onPress` of the ZulipButton needs to be wrapped in
curly brackets.  The previous notation returned the result of
the `dispatch` call, which is an Action.  This return type is
incompatible with the return types allowed by Flow.  Flow hasn't
previously complained because the type checking didn't work
properly with `connectWithActions`.