version: Bump version to 13.5.89, an Android-only stable release.
This release contains a workaround (#2595) for an Android-only issue
in React Native upstream -- which itself triggers an iOS-only issue
also in React Native upstream.

So to thread this needle, we're releasing the workaround in an
Android-only release.

If we end up maintaining any kind of platform split like this for more
than a version or two, we'll of course want to do it with conditionals
in a single source tree rather than with different versions of our
codebase.