compose box: Use 'controlled' inputs on iOS
This reverts the `ComposeBox.js` file to its version before making
it uncontrolled. Keeping it as a separate copy-pasted file for
simplicity.

This is a temporary, but needed because of a bug that does not allow
us to reset an uncontrolled input's value.

A likely fix, not yet merged is here:
https://github.com/facebook/react-native/pull/18278