flow: Define a base ServerEvent type.
Server events contain an `id` property that is not really used by the
reducers (it is a incremental id set by the server)

Most of our event actions already contain this property, because they
only add to the server event object.  Base all `Event*` types that we
can on `ServerEvent`, and remove `id` from the few individual types
where we've written it already.