webview: Add emoji styles as static `emoji.css`
Instead of generating the data dynamically (in cssEmojis.js) just
store it in `emoji.css`.

This makes the dynamically generated parts much shorter and
easier to debug.

We keep the `cssEmoji.js` file intact, as we can use it to generate
the `emoji.css` file again if input data changes.