webview: Load JS files from assets
Instead of loading imported-strings-as-js we just refer to the
asset files of the device.

Also, move the `scripts` code to the bottom of the page.
This was initially moved to the top so the reported error code-lines
are more predictable (the different html was messing up the count).

Now we don't need that, so simplify.