flow: Properly annotate EventUpdateMessageAction.
EventUpdateMessageAction was the last `any` action that was part of
the union type Action.  Previously, Action has always included at
least one specific action type that hadn't been properly typed, that
is, had been set to `any`.  This allows Flow to just replace Action
with `any` whenever it was used in a way that would have caused an
error, making Action worthless.

For now, leave a $FlowFixMe to preserve that status quo explicitly,
because there are a few errors we get when we remove it.  We'll fix
those in followup commits.