notif: Don't try to scroll to message's position.
On Android (though not on iOS), when following a notification we
would attempt to position the message list at the message the
notification was from.

When running the app from a 'suspended' or 'not running' state this
does not work correctly, as we are not anticipating that case in
`getMessageUpdateStrategy`.

Luckily, the fix is simple -- if there's a long conversation which
the user was e.g. notified in the middle of, it's better to take
them to the first unread message anyway, rather than the one that
generated the notification.  So just do that, by simply not passing
a message ID to `doNarrow`.

Fixes #2699.