webview: Transfer css rules from .emoji to .emoji:before.
Some background: the css ruleset `.emoji` styles the plaintext
that represents an emoji. This plaintext is, by default, hidden
(by setting the color to transparent). The visual emoji icons are
inserted via the `emoji:before` selector. Move the declarations for
height, width, and vertical alignment to the `emoji:before` rule-set,
as (I believe) they are intended to style the icon (and not the
plaintext).

Fixes #2693.