flow: Remove disused *ActionCreator types.
These were used as the properties of the now-removed Actions type.
They're basically each the type of some function in our codebase, and
those functions should all have the type signatures right there at
their definitions; so it's much better not to duplicate them all here
if we can avoid it.  And now that Actions is gone, we can!

One such type remains: SaveTokenPushActionCreator has a few references
still, so it survives for now.