Move action button to AvatarList
The floating action button to create the group/invite users in UserPickerCard was aligned to be on the bottom of the scroll view.
This caused problems when the list was large as the user had to scroll all the way to the bottom to find the button.
The button is now moved to the AvatarList component which appears at the top and appears on the right end of this component.