ui: Use cached data for conversations tab.
Fixes #2725

Previously we incorrectly showed a loading spinner if users data
was loading regardless of wether we had user data already.

For better user experience and to be consistent with the rest of
the app's behavior, show the cached data while updating the user
data. Show spinner only if no data is cached.