pms: Rename ConversationGroup to GroupPmConversationItem.
Historically we've tried to reserve "conversation" as a term for
specifically a PM conversation.  But when people are talking with
each other on a stream, that's also definitely a "conversation" by
the normal English meaning of the term; and in the user help, the
web UI, and the webapp's code, we use it in that more general sense
too.  So this doesn't work as a jargon definition that will be
clear to the reader.

Instead, for the set of PMs with a given other user or set of users,
we'll say "PM conversation".  This is one of a series of commits to
make that change.

Also ConversationGroup isn't really a group; it describes group PMs.

[greg: expanded commit message]