compose: Factor out "clear input" logic into `updateTextInput`.
This will let us re-use the same logic for the topic input, and for
initializing the inputs on mount when we make them uncontrolled.
While we're here, add some comments explaining the trickier bits
of this logic.

We also add a `setMessageInputValue` wrapper that also makes sure to
call the event handler `handleMessageChange`, for consistency in
behavior with what gets called when a user types text themselves.

The new function adds a call to `setNativeProps`.  This doesn't have
any effect as long as the input is controlled, and it makes a simpler
transition to the uncontrolled version coming next.

The way controlled/uncontrolled state works in TextInput is:

  When controlled (and without this `setNativeProps`):
   * we set message in state
   * invisibly to us, RN calls setNativeProps to reconcile the
     difference between what value we want and what is the current
     value of the native component

  When uncontrolled:
   * we manually call setNativeProps
   * we also set state for `message`, but that is only so we track
     the current value of the input in JS-land

  In this commit:
   * we call setNativeProps with value X
   * we also set state to X
   * React Native invisibly to us checks if desired value differs
     from what the value of the native component is.  X === X so
     RN does not do anything.