android: Make release and debug app build coexist in the same phone.
Now when we run `react-native run-android` will install debug
build of the app. But it will try to launch the app with the
default package name (without any suffix), as it is launched by RN.
We need to tell RN to launch debug build which has suffix 'debug'.

So use `react-native run-android --appIdSuffix=debug`.

Fix:#2705