android: Use different app icon for release & debug mode.
Now both the app (downloaded from play store & build from development
env) can coexit in the same phone. So to distinguish them, it's better
to have different launcher app icon for both of this.

Debug app icon is same as release one but with a DEBUG label on the icon.