nav: Get streamId param from props instead of from state.
Currently, in screens where there's a stream ID in the route params,
we're using a selector to get it from the global Redux state.  The
standard practice would instead be to get it from the `navigation`
prop that was actually passed to this screen component, and the way
we are doing it causes problems.

For example: when popping the route off the navigation stack, the
navigation params in the redux store change to null instantly, but it
takes time to unmount the component. This causes the screen to be
repainted with `NULL_STREAM`, which looks half-blank and broken.

Instead, use the prop.

This fixes part of #2746.