a11y: Make avatar in webview voice-readable
Adds an `alt` prop to the `img` tag showing the avatar inside
we WebView message list.

Previously a voice-over assistent was reading the item with some
random numeric number making it very unusable. Now it reads
"John Doe graphic" where "John Doe" is the name of the person.

This is consistent with how all other apps treat this UI element.