a11y: Make Touchable accessible
Add an optional `accessibilityLabel` property to the Touchable
component.

By default all "touchable" components in RN have `accessible=true`,
i.e. are accessibility components:
  https://facebook.github.io/react-native/docs/accessibility
so the same is true of our Touchable when we actually make it
touchable.  In that case, we just pass the label through.

If we render as a plain `View`, then pass the label through, and also
set `accessible=true` just if there was a label.

If we have text components inside the `Touchable`, often the OS will
figure out a good label for the component.  Sometimes though it either
doesn't or we can provide a better one.