a11y: Remove `onPress=nullFunction` from Avatar components
Any `Touchable` is accessible by default. Any `Touchable` that
does not have an `onPress` though is rendered as `View` which can
be accessible but is not so automatically.

We do not want all avatars to be accessible. To the average user
an avatar with onPress of `nullFunction` is exactly the same as
one with none. To a person using voice-over there is a difference.

By removing the default onPress value entirely we make sure these
components are not accessible by default. Thus many elements that
were misleading or 'fake accessible' are now ignored (meaning that
the voice-over function will try to read them but we don't want to)

Such components are:
 * PM titlebar's avatar
 * Group titlebar avatars
 * any other unanticipaded component