notifications: Merge notifications.* into one file
Flow does not recognize `*.android.js` and `*.ios.js` files for
what they are. Also, sometimes they are more trouble to maintain
than it is worth it.

This commit merges the two platform-specific files into a single
one. This is straightforward and mostly consists of `if`-ing the
`Platform.OS` value.

Also turn off an ESLint rule that insists on the separate-files
approach.