apiFetch: Handle unauthorized api call.
Dispatch logout action when response status is 401.

If api key is changed on the server, and we still
uses old api key to fetch data. App ends up in loading
forever.

This PR dispatches logout action with a toast message to login again.
So that when user login again app can store new api key.

Fix: #2770