diagnostics: Re-add timer for store hydration.
The store hydration timer got removed in bf659a25c when we upgraded
from redux-persist v4 to v5 due to structural differences in the redux
initialization. Re-add a timer. Do this with a new, reusable Timer
component that logs the duration of its existence to timing.js. This
works for roughly measuring the redux-persist execution time, because
that's all the app is doing while the loading screen is shown. No
other big actions distort this measurement, because no significant
actions are dispatched as long as the state is not properly loaded.