diagnostics: Fix timestamps and durations.
The previous now() used performance.now() if available.  This led to wrong
calculations of the timestamps in TimeItem, because performance.now() does
not return the number of milliseconds since 1970, which is what we need to
calculate the timestamp.  There is no documentation on the return value for
RN, but from empirical observations that value is not the number of
milliseconds since 1970.  (In the browser, it returns the milliseconds
since the page load.)

Furthermore, performance.now() does not actually yield higher
precision anymore, because browsers add up to a millisecond of noise
to mitigate Spectre:
  https://developer.mozilla.org/en-US/docs/Web/API/Performance/now

Also, don't divide Date.now() by 1000.  The format() function from
`date-fns` that we use for formatting the date number to a date string
seems to interpret the number argument as milliseconds.  (Unfortunately,
there is no documentation on the unit type to confirm this.)