ModalSearchNavBar: Remove redundant code
Earlier ModelNavBar made search functionality available depending on
props passed to it. Later this was spun off into two components,
ModalNavBar and ModalSearchNavBar, with ModalSearchNavBar being used
when search was required. There was some leftover logic from when one
component provided the search functionality in ModalSearchNavBar,
which has been removed in this commit.

The value of isSearchActive is always true for ModelSearchNavBar,
since the prop searchBar is never passed by anything in the code
anymore. The only function that can set it to false,
disableSearchActiveState() is never called because that requires
clearSearchInput prop to be true, and that is never passed by anything
as well. This causes some code to be never exectued.