android manifest: Add debug package name in comment.
Why this is required?
In continuation of making release & debug build co-exist in the device
at the same time, package name suffix is mentioned in the build.gradle
(app).
By default RN reads package name from manifest via parsing XML, and
tries to launch app whose package name is associated with that.
So with this pretend that our package name is same as with the debug suffix
i.e `com.zulipmobile.debug` so that RN lauches this debug build app after
installing (via `react-native run-android`).

As it is added in comment, so no effect in the build process.
Only it affects RN which parse this file as a string and extract
package name.
https://github.com/facebook/react-native/blob/46ffb106278425631c26049fa50b185e90789c9c/local-cli/runAndroid/runAndroid.js#L111

Without this, RN will take `com.zulipmobile` as package name
from manifest and tries to launch this, which is the package name
for release build.

Note: We have added package name suffix in the build.gradle for debug
build so on running `react-native run-android`, it installs app
with package name as `com.zulipmobile.debug` (as by defualt variant is `debug`).
https://github.com/facebook/react-native/blob/46ffb106278425631c26049fa50b185e90789c9c/local-cli/runAndroid/runAndroid.js#L439

Another way could be sending parameter in the `react-native run-android`
like `react-native run-android --appIdSuffix=debug`.
This approach is not good because it is breaking
`react-native run-android`, which is the traditional way of running RN apps
on Android.