android: Create a entry point activity for app.
Which just extends ZulipActvitiy, which also contains
all the logic.

Why this is required?
We are pretending that our package name is `com.zulipmobile.debug`
by adding a comment in the manifest (just for RN).
(Note: Read previous commits for more info)

If we are pretending that our package name is something different
we want valid activity (entry point activity) which should be
present at this pretending package.

Thus extracting logic will help in avoiding logic duplicacy.