title: Make `getTitle*Color` selectors independent of nav state.
Fixes #2797.  Fixes #3139.

The problem: Let's take an example, user is in stream narrow.  Now
they open the topic screen (by selecting topic icon in the nav bar).
As navigation state is integrated with our Redux global state, it
changes nav state, and now currentRoute is no longer `chat`.  Thus
changing the output of `getIsInTopicOrStreamNarrow`, and this change
gets propagated to `getTitleBackgroundColor` and finally to the
component -- changing the title colors for the existing chat screen.

Instead of using `getIsInTopicOrStreamNarrow` which looks at the nav
state, just consult the `narrow` parameter we're passed, which comes
from the component's props.  Now the chat screen keeps its own color
for whenever it might be visible again.

Also clarify the tests a bit.