redux-persist: Add compression, on Android.
The motivation that makes this important to do now is the storage
limits imposed by RN's AsyncStorage on Android.

In particular, on Android the open-source RN uses a backend called
AsyncSQLiteDBStorage (it prefers one called AsyncRocksDBStorage if
present, but the implementation of that one seems to be internal to
Facebook).  In that backend:

 * A limit of 6 MiB is imposed on the size of the entire database.
   This is a choice made inside RN; see `mMaximumDatabaseSize` in
   upstream's ReactDatabaseSupplier.java .

 * The system's SQLite library retrieves data in windows of, by
   default, 2 MiB, and rows larger than that will cause an error.
   The backend stores each key-value pair as a row, and redux-persist
   gives it each top-level subtree of the state into a key-value pair;
   so when one top-level subtree goes over 2 MB or so, it fails,
   discarding the whole persisted store:
     https://github.com/rt2zz/redux-persist/issues/284

   The `users` state subtree for chat.zulip.org now comes to a little
   over 1.8M characters, so we must be approaching this limit.  (This
   also means it's getting encoded as bytes reasonably efficiently,
   presumably as UTF-8.)

   Looking at Android docs, it looks like a sufficiently determined
   Android app could use `SQLiteCursor#setWindow` to change this size;
   but the code that'd have to do that is deep inside RN, so even
   doing that as a quick hack would take some work.

On iOS, by contrast, AsyncStorage uses a simple custom key-value store
with large values stored out of line (see RCTAsyncLocalStorage.m .)
That one doesn't seem to have any such limitations.

A library "redux-persist-transform-compress" is intended to solve this
problem; but it implements the compression algorithm in JS, which
makes it slow and adds serious lag to the app.  So we don't use that.

Instead, implement a native module to do the compression in Java, and
provide a wrapper for RN's AsyncStorage to use it.  Note that we
can't do this with redux-persist's "transformer" concept, because
transformers don't support async operations.  Loading and compressing
state can take enough time to cause frame drops, so we need it to
happen async.

[greg: expanded the explanation / research about storage limits;
 also deactivated the actual invocation of CompressedAsyncStorage,
 pending the needed fallback logic.]