Animate "No internet connection" banner.
Expand and collapse <OfflineNotice/>, similar to Facebook Messenger.
Add a small delay so that the OfflineNotice is not immediately
displayed when e.g. narrowing to a topic.  This makes its appearance
more prominent to the user.

We don't use the native animation driver, because it doesn't support
animating layout properties like `height`.