async: Loosen a racy test to try to prevent flakes.
This test failed in a CI build on #2729:
  https://travis-ci.org/zulip/zulip-mobile/builds/405727850

Probably the system was busy and it took a bit longer to get to this
task than the minimum.  That's perfectly normal behavior, and we
shouldn't fail on it.

It'd be nice to spruce this up so it e.g. reports the actual duration
if it fails, but this quick fix should do.