eslint: Disable rule jest/no-disabled-tests.
This rule is set to warn by default, and it fires on any test marked
to be skipped, as with `test.skip`.

That basically defeats the use of the "skipped test" feature, which is
a useful tool.  Disable the rule.

If we could make it so that a normal `yarn test` would show warnings
only on *modified* files -- and, crucially, warnings on modified files
showed up in CI, so they'd be visible in code review -- then having
this rule warn would be exactly right.  But we don't have that
functionality, so we have to keep the codebase clean of warnings as
well as errors in order to keep the linter useful.