Created getAllRealmEmojiByName and getAllActiveRealmEmojiByName
These selectors return an object that has the structure expected by
the rest of the application.

Previously the selectors getAllRealmEmoji and getActiveRealmEmoji
returned objects that used the emoji ID as the key for each object
that contained that emoji's details. All the code in the application
that dealt with realmEmojis expected the emoji name to be the object
key. This caused realm emoji to be not displayed in the messages if
someone had reacted with one from the web app, not show up in emoji
auto complete and the new EmojiPickerScreen.

These new selectors that use the emoji names as keys, shall fix all
of these issues.