Upgrade redux-persist from 4.10.2 to 5.10.0
redux-persist v5 introduces a couple breaking changes. These are
described in detail in
https://github.com/rt2zz/redux-persist/blob/master/docs/MigrationGuide-v5.md

Here are some of the changes that affect our code (other changes have
already been addressed in the previous commits):

* v5 auto-rehydrates the state by default, so we don't need to call
autoRehydrate() explicitly anymore.

* v5 dispatches different actions during the first rehydration. This
makes it incompatible with `redux-action-buffer`, because
redux-action-buffer silently buffers all these actions which prevents
the redux-dispatch init from completing. Conveniently, v5 also ships
with a new component, `PersistGate`, that buffers all
non-redux-persist actions and can also display our loading screen
while the state is getting initializeed.  Therefore, get rid of
`redux-action-buffer` and replace it with `PersistGate`.

* v5 uses a different storage format that is incompatible with the old
format. Automatically convert the old format to the new format with
getStoredStateMigrateV4().

* v5 has internal Flow errors, so exclude the module from
flow and add a type stub with flow-typed.

The latest official release of redux-persist is v5.7.0, and the latest
version is v5.10.0. The code changes between v5.7.0 and v5.10.0 are
mostly bugfixes and no bigger features, so use v5.10.0.