notifications: Handle GCM to remove messages from the notification
Requires the message id of the message to be removed and event called
remove. The current data structure saves the list of messages grouped
by their narrow.
There can be two cases, the notification to be removed can be either the

Last one - Remove the full conversation as if the last message is read
then the previous ones are read also.
Before the last message - Remove that particular message from the list