persist: Switch on redux-persist-migrate.
The previous commits checked the code of redux-persist-migrate into
our codebase, but didn't use it anywhere.

Plug redux-persist-migrate into redux and redux-persist, so that
anyone can straightforwardly add migrations.

To know what migrations have been applied, redux-persist-migrate
stores a `version` object in the redux state.  We can configure under
which top-level subtree that key should be stored.  We create a new
top-level subtree `migrations` to store the `version` key, so it is
isolated and doesn't interfere with other data.