presence: Improve focus ping
Pass `new_user_input` as `true` instead of `false`.
I had incorrectly assumed before that this has something to do
with the user putting in new 'input' (typing)

The ditermination of wether it should be `true` or `false` is made
on the web if the user has actually interacted with the app.

On mobile we do not have a long running browser tab that might be
in the background, thus we should pass `true` always.

As discussed here:
https://chat.zulip.org/#narrow/stream/3-backend/topic/presence.20update.20fields