state: Clarify logic to reset unread states.
The current behavior of all unread states when receiving a
realmInit response with no data is completely correct.

Without changing this behavior, replace the value used to reset
it from `NULL_ARRAY` to `initialState`. This makes it more obvious
what our intent is and a bit more consistent with the other reducers.