state: On missing `mute` data on realm_init reset state
The current logic to change or not the mute reducer on realm init comes from:
https://github.com/zulip/zulip-mobile/commit/d91dc06796f73b02e0956ad07fdb22ce83673106#diff-b16ebc34e88ab6ad4461b6c019e2de0d

It is preventing repaints by doing deep equal comparisons.
While this is a very useful technique to optimize performance,
we consider this an anti-pattern if used generally everywhere.

To remove that pattern and to make the realm_init consistent with
the other reducers reset to `initialState` if no data provided.