android notif: Refactor slightly to push logic into onReceived.
This is a pure refactor, though the logic around this code is a
bit tricky.

This `getPushNotification` method gets called from one spot in the
wix/react-native-notifications library, which in turn gets called in
three places.

  * One is for local notifications, which we never do.

  * One is from the library's `GcmMessageHandlerService`, when a GCM
    message arrives.  At that call site, the next thing that happens
    is the return value's `onReceived` method gets invoked; then the
    return values are all thrown away and that's the end of the
    handler.

  * One is from the library's `ProxyService`, which receives things
    like the intent from the user pressing a notification.  On just
    pressing the notification, this method is called with a bundle
    with the same props as the original GCM message; on choosing the
    "Clear" action, it's the one-key bundle we created in
    GCMPushNotifications#getNotificationBuilder.  In any case, what
    happens next is `onOpened` gets invoked on the return value if
    non-null, then that's it.

In the third case, it's actually wrong that we're adding to
`conversations` with `addConversationToMap`, though it doesn't much
matter because the `onOpened` then clears `conversations`.  So it's
good to remove that call from that case.

In the second case, we can freely move logic into `onReceived` with
no effect on behavior.  And this seems like the better spot for this
logic, because it only really makes sense in this case and that's
what `onReceived` is for.