android notif: Ignore GCM messages we don't understand.
The GCM messages sent by the Zulip server use the key `event`
(think "event type") to identify the type of the message.  In all
current versions of the Zulip server, the only value for `event` is
`message`, meaning the GCM message is describing a Zulip message
which the app should notify the user about.

If we get a GCM message with some other event type, then we can't
possibly interpret it sensibly.  If we try to do so, then it's
impossible to introduce new features requiring new event types
until all users have updated the app to a version new enough to
know about them, because older versions are likely to produce
gobbledygook to show to the user.

For example, server commit bc43eefbf, which will be in Zulip
Server 1.9 but with its behavior disabled by a setting, introduces
an event type of `remove` -- meaning the GCM message is about a
Zulip message which the user has now read, and we should *remove*
from any notifications display.

So, ignore such things.  The user simply won't get the benefit of
the new feature until they upgrade, like other new features.