andorid notif: Implement GCM token refresher services
Generates GCM token using the sender ID stored in the AndroidManifest
using an IntentService called GCMIdRefreshHandler.
An additional service extended by InstanceIDListenerService is used to
call the intent when the token is automatically updated by the google
servers.

In the JS side a function called refreshToken can be called via tha
GcmTokenRefreshReactModule module and when the refresh process is
completed, it is emitted into a notificationGcmTokenRefreshed event
which can be collected in the JS module