android: Update `buildToolsVersion` to `27.0.3`.
Plan is to update gradle plugin to 2.3.+ to get advantage of instant
run. Which minimal requires `buildToolsVersion` as `25.0.0`.

Also `buildToolsVersion` is the version of the compilers (aapt, dx,
renderscript compiler, etc...) that you want to use. For each API
level (starting with 18), there is a matching .0.0 version.

https://stackoverflow.com/a/24523113/5612089

So it's better to compile app with latest version, to make sure it
works on newer android versions.