action-sheet: Update to v1.1.0 of Expo's action-sheet library.
This is the library that provides an Android equivalent for RN
upstream's ActionSheetIOS; v1.1.0 comes with a bunch of changes to
the layout, intended to make it conform to Material Design.

The new layout isn't great, but it does happen to fix the long-pending
issue of text in options being cut off on One Plus and Oppo phones.
(Perhaps because it no longer makes the text bold.)  Screenshots
in #2910.

Fixes #2365.