app_refresh: Change how `typing` state resets
Resetting `typing` state on `REALM_INIT` does not make sense so
remove that. Resetting this on `APP_REFRESH` though makes a lot
more sense, do that instead.

After 10 mins or more we can be sure the typing state did change
and resetting to the default one (no one is typing) is a safe option.