Improve autocomplete popup sizing
Fixes #2917

Several previous changes regressed the styling of the Popup.

The ideal size of the popup can be described as:
'the whole screen sans the keyboard and with some padding around'.

The lack of CSS `position` prop value of `fixed` being supported,
and the lack of React portals (another option) support makes implementing
this tricky and requires a lot of rework.

This commit finds a good middle ground - use a max height of half the
screen height minus the compose box's height.