Changed type of params in NavigationState
It was known that params was on object I failed to understand what it contained so i ended up using Object as the type. On further study, i realized it only contained only a single field, that was "narrow". 
Again type of narrow was unknown. The type of narrow was determined after going through src/message/fetchActions.js where the type of narrow was specified as Narrow.