api: Skip `Authorization` header when no `auth` provided.
Several API calls do not require authentication. We currently do
a simple `fetch` for them and skip our custom function.

There is a benefit to using our functions though. For example
providing a userAgent key.

This commit inspects the `auth` object, if it has an empty
`apiKey` (empty auth object will work too) it does not include
the `Authorization` header in the request.

Also adds a test to confirm this works as intended.