Do not discard but store `flags` data
Improves but does not completely resolve #2940

On app's start there is a potential (very likely) mismatch between
the stored messages and the flags associated with them (read, starred)

The first issue is that we do not update the messages themselves
but we do update the flags state. But we do not preserve the state!

The second issue is that the flags might have changed since the last
run, again very likely, especially for the `read` flag if the user
have used the web app.

This commit resolves the first issue. The second one requires more
work, mostly the code implementing short/long-lived queues.