autocomplete: Return null if filter length is zero.
If filter string is not available, then return null from component and
render nothing in autocomplete popup. Because we suggest matching
items based on filter string (user have entered after ;, @ or #) in
the popup. So if filter string is empty, then we don't have enough
info to suggest.

For example: if user is typing `some text # some more text.....` Then
as here filter string is empty, so save re-render of animation
component.

partial fix for #2791.