autocomplete: Fix sudden flash just after entering prefix.
What is causing issue?

`AutocompleteView` is dependent on `text` and `selection` (the only
props). Ideally what one should think is, both will change at the same
time. But it is not the case (might be issue in upstream). So whenever
new char is entered or removed then both `text` and `selection` are
changed. But the twist here is both doesn't changes at the same time,
first `AutocompleteView` is rendered with new `selection` value and
then `AutocompleteView` is again called with new updated value of
`text`. (How to verify this? `console.log` is your friend here).

Now when `#` is entered, `AutocompleteView` is called with `text` as
'' and selection as `start:1, end:1`. As the `text` is empty string,
corresponding `AutocompleteComponent` can't be found. But with current
code, parent `AnimatedScaleComponent` will be rendered with empty
child. This change also saves this extra redundant render.

Now for the second re-render where props are `text` as `#` and
`selection` as `start:1, end:1`, `AutocompleteComponent` can be
determined and `AnimatedScaleComponent` will be re-rendered. At this
point filter is '', i.e empty string, which matches all the
items (with current logic of filtering items in people, emoji, stream
autocomplete). Thus full list is rendered and visible for a monent.

Now `visble` props of `AnimatedScaleComponent` comes into the picture,
as the filter length is 0, it is false. Thus `AnimatedScaleComponent`
realizes that it's display style should be `none`. Why this delay in
realization? Because `visible` is taken care in `animatedStyle`, and it
obvious that animation styles are applied after component gets
rendered, then only any type of animation will be visible. Thus now
`AnimatedScaleComponent` makes it's display `none` and leaves behind a
sudden flash of full list.

How it got fixed?
If `lastWordPrefix` is something for we are not interested (i.e
neither @, : nor #) then return null from `AutocompleteView` instead
of rendering empty animation component.

Fix: #2791