animatedScaleComponent: Add optional prop `display`.
If children of animated component is not visible then do not display
animated component. This dependency was introduced in
9c1dea201d4920ae04c00e67cba1d16afaf85d76 and
7fbe5bd636b0b3b037c2bec32b31b0ed4a9724ca.

Initally this component was used by unread notice. So if there are no
unread messages then we don't want this, thus taking no space on
screen (i.e display none).

But autocomplete floats/overlay on the screen. So display none is not
a good option in this case. So make `AnimatedScaleComponent` more
customizable.

Read more display prop at
https://facebook.github.io/react-native/docs/layout-props#display.