autocomplete: If filter string is empty then return null.
If filter string is empty then it matches all the possible items, so
it starts rendering all the items in the screen. But in
`AutocompleteView`, animated component is only visible if filter
string is not empty. So it will start scaling down it to 0. Thus
resulting in splash of popup becoming almost full screen (or to the
max height) before disappearing. If this condition is applied in the
autocompleteView then it will directly unmount component, without
animation. Also we want a inital render of animation component (i.e
with empty filter string), as then onwards it will go in
componentWillReceiveProps (in animation component) and start
animation. As componentWillReceiveProps is not called for iniatl
render. And we will miss in animation.

Fix: #2791