Lightbox spinner: Show spinner in Lightbox when slow to load
Previously, Lightbox remains black until an image is loaded from the
server.

Spinner will be shown to user in Lightbox, while an image is being
loaded. New state 'loaded' is defined in Lightbox component. When an
image is loaded, the state will be changed from false to true and display
value of 'PhotoView' also will be changed from 'none' to 'flex'.

Fixes #2896.