state: Do not rehydrate 'flags' state
Previous commit makes sure we store the `flags` state.

This is better than trying to restore flags from `messages` as we
do not mutate this data. So any state that was changed during the
previous run of the app would not get restored properly.

The simpler solution of storing the `flags` is far superior to this
approach while not being a complete solution.