redux: Add `disableNavTransitionEffect` param to `connectPreserveOnBackOption`.
This is a optional param to decide whether `isNavTransitionGoingOn`
should be considered or not in evaluating result.

As we don't want to consider `isNavTransitionGoingOn` while evalaution
for `TopicListScreen`, because as the screen in pushed, it performs a
api call whose result is saved in global state. So to take care of
this changes, components wants a re-render. And there might be cases
where global state is updated while transition was taking place and
thus avoiding re-render.

As `TopicListScreen` is not depended on global nav state (which
changes while transition takes place), so this will not have any side
effects/extra re-render.

Fix: #2808