migrations: Explicitly remove `messages` subtree when dropping it.
This has no effect right now, because there's no `messages` top-level
reducer and so redux-persist won't even look at `messages` in the
persisted store and will pass the migrations a state that already
lacks it.  That was our reasoning in leaving this out when we added
this migration in 638cddce0 "state: Rename `messages` to `narrows`."

But now we're about to create a new, differently-structured subtree
under the same old name `messages` (in an upcoming change to normalize
the `narrows` state.)  When a user upgrades across both these changes
together, the old reasoning won't apply.  So we should make the
removal explicit after all.

In retrospect, the right practice is to include this logic in the
first place.  There may be other times when we remove a top-level
subtree and then much later reuse the name for a new one, when the
removal is less fresh in our minds; removing it explicitly here in the
migration makes things robust.