flow: Fix `children` type errors in two components
Make `children` of type `any` for `Centerer` and `ComponentList`.
Current version of Flow is too strict and the `ChildrenArray<>` type
is not working - and it should. Flow is trying to guess between
two almost similar types. Trying to fix this in a better way seems
to be way more complicated (tried for an hour) and it might get
changed in a next version of Flow.

Making this `any` for now seems like a rather sensible compromise.