stream: Eliminate NULL_STREAM.
See reasoning on the recent similar commit for NULL_MESSAGE.

In `getStreamFromId`, there's no sensible default value and the
right behavior depends on the caller's needs.  It turns out that
every call site is a line just like this, in a `connect` call:

  stream: getStreamFromId(props.navigation.state.params.streamId)(state),

In that situation, on the one hand there's nothing we can do that
will avoid the screen looking quite broken; and on the other, for us
to get here with a stream ID that doesn't belong to any stream we
actually know about should be a highly avoidable kind of programming
error.

Moreover, checking the call sites of each navigation action creator
that passes a stream ID in the nav state (thanks, VS Code, for making
that efficient -- Shift-F12 on each one in turn, in navActions.js),
the argument in every one of them looks like `stream.stream_id`, which
should mean we don't currently have this highly avoidable kind of bug.

So, just throw an exception.

(There was a time that we used this function's default return value as
the default values when creating a stream, so that the way to change
those defaults was to edit NULL_STREAM -- yikes!  See 94ce54dba.
Fortunately we fixed that soon after, in f9085d3f7.)

If we had some facility for rendering an "oops!" screen on error,
this would be a good place to use it just to be sure, even though
the risk is low.