Upgrade to React Native 0.57.1
(The currently latest version of 0.57.2 has issues)

Both updrade tools failed to run:
 * react-native-git-upgrade
 * react-native upgrade

This will be a manual but thorough old-school upgrade.
Not a problem, since we did that already for all but the last upgrade.

First step is just to upgrade the dependencies and the flow config file.

We also replace "babel-preset-react-native" with "metro-react-native-babel-preset"
and then .babelrc changes to reflect that.