rn-cli: Remove obsolete Metro blacklist entry.
Since RN v0.55 (with RN commit 54dc11a5f), rn-cli has had its own
default blacklist consisting of an improved version of this pattern,
so this isn't necessary.

With RN v0.57 (commit a32620dc3), the format of this config file
apparently changes, and this breaks.  Take that as a reminder to
just delete it.

[greg: expanded explanation]