docs/arch/realtime: Write up a design doc on the events system.
This is a client-focused companion to the existing subsystem doc:
  https://zulip.readthedocs.io/en/latest/subsystems/events-system.html
which is focused primarily on the server side.

The design described here is the product of some time I spent just
today thinking about how to effectively structure a client to use the
architecture of the Zulip events system.  It represents a direction I
believe we should go, though it's not a finished document.