lint: Fix several issues but disable `react/destructuring-assignment`
The rule `react/destructuring-assignment` is helpful by suggesting
descructuring inconsistencies. It is not perfect though. Several of
the suggestions are needed (initializing class props) and several
others are better left as-is (when comparing this.props and nextProps)

So, fix a lot of the issues reported to achieve consistency, but
still disable the rule.