flow: Give a type to the containing function in `Input`
With the Flow 0.75 upgrade that comes with RN 0.56, the autogenerated
`flow-typed` stub for `react-intl` sees two possible types for this
`children` prop:
 | ((...formattedMessage: Array<React$Node>) => React$Node)
 | (string => React$Node)

where previously it saw only the former.  But then it complains
because it can't decide which to use.  Giving the concrete type
fixes this confusion.

Until the upgrade, on the other hand, Flow doesn't realize simply
`string` is possible here; so we need a fixme comment.