flow: Change Style type to `any`
React Native have improved their internal Flow types significantly.
They have a lot of concrete style types and they conflicted with
what we had (a decent approximation)

I tried for hours to implement something sensible that kept our
styling types concrete and helpful.

The best approach looked like this:

```js
import type {
  ImageStyleProp,
  TextStyleProp,
  ViewStyleProp,
} from 'react-native/Libraries/StyleSheet/StyleSheet';

type Style = ViewStyleProp | TextStyleProp | ImageStyleProp;
```

This confused Flow as it wasn't able to choose between the three
when two or more were viable.

Trying to specify one of the three types directly instead of Style
looked like the right option. After coding this for a while though,
it was obvious that it required too much change, and the benefit per
effort was not good.

We should keep an eye on how the RN internal typings and Flow itself
progresses and reconsider using more concrete types in future.