flow: Add $FlowFixMe comments for `ZulipStatusBar`
A rather strange issue arises from the `flow-typed` definitions
and the way Flow inferes the return type of the ZulipStatusBar
when wrapped with `connect`.

A rather long investigation and search for a concrete fix was futile.
There appears to be nothing wrong with our code and I suspect this
will be fixed either with a newer Flow version or with a change to
the type definitions in `flow-typed`.