image-picker: Fix iOS camera flow by waiting for image to be saved.
On iOS, if you use the "camera" icon in the compose menu to try to
take a photo and send it, it doesn't work -- no message actually
gets sent.

Looking with the React Native Remote Debugger, this manifests as a
console warning that `String.prototype.split` was being invoked on an
undefined value in `getFileExtension`.  On investigation, there was no
`fileName` property on the `response` object passed to our callback
`handleImagePickerResponse` by the react-native-image-picker library.

Turns out that as the documentation says:
  https://github.com/react-community/react-native-image-picker#options
there will be a `fileName` only if these two options are set to true.
So, do that.

The first time through the resulting flow is still cumbersome with 3
permissions modals, and doesn't actually work -- but at least every
later time now works, which is progress.  See #2973 for followup.