createStreamScreen: Add `keyboardShouldPersistTaps="always"` to Screen.
Let's take an example, user just pushed create stream screen, now as
the details/fields are empty, button is disabled and on clicking
button keybaord is dismissed. Which is totally unexpected.

Basically as the button is disabled, so onPress is undefined and View
is rendered in the Touchable. So this touch gets propogated to
ScrollView, which dismisses the keyboard.

This improves the user exprience.