flow: Update .flowconfig
Because of the removal of @providesModule in React Native 0.56 a
confiuration in Flow was changed. Both the removal and the config
is implemented in this PR:
https://github.com/facebook/react-native/commit/d5e9e55fa

It configures name reducers which are series of replacements that
are applied in order to reduce a filename to a module id.

For more information on haste name reducers here:
https://github.com/facebook/flow/commit/ca4a3d9f4603f0636a6efa79cb222525d1a67497