webView: Make `getStartAndEndNodes` more concrete.
Before: A attempt was made to find last message visible on the screen
(with current scroll position) at a specific cordinate i.e (200,
window.innerHeight - 20). Now if list is short then, this node will
not be a message node and 0 will be passed, which will not not make
any sense as start > end (see `msg.id >= fromId && msg.id <= toId` in
`filterUnreadMessagesInRange`).

After: If above mentioned attempt is failed, then get all messages
node and get id of first & last message node and pass it to react
component for further use.

Fix: #2988