android: Replace default icon's shape. Add round icon.
Since Android 7.0 the launcher supports rounded icons.
The default icon should be square with rounded corners.

Guidelines here:
https://material.io/design/iconography/#grid-keyline-shapes

This replaces the default round icon with a more rectangular
default one and an optional round one. The round one will be
used only if that is configured in the phone's launcher.

Custom icons versions were added for or Debug build too.

Added new icons for:
 * release build, default icon
 * release build, rounded icon
 * debug build, default icon
 * debug build, rounded icon

Added a new high-res icon option in the folders named `xxxhdpi`.

Generated with the help of Android Asset Studio:
http://romannurik.github.io/AndroidAssetStudio/