compose: Remove unused drafts-related code
Using `componentWillReceiveProps` the way we did should not be
considered a best practice, and should not be used if better
options are available (there are other lifecycle events too).

We still need the `editMessage` comparison, though future rework
will make it not needed.

We don't need the `narrow` comparison though. It can never be true.
It used to be usable before, not anymore.

We get the prop down from several components. The hierarchy is:
ChatScreen > Chat > ComposeBox

ChatScreen gets the `narrow` value from the React Navigation's
`navigation.state.params`. This value does change only when a new
screen is created. So the same instance of the same screen will
never receive a different value. This comparison will always produce
`false`. Removing it!

This was confirmed both theoretically by analysing the code thoroughly
and empirically by putting a `console.log` and navigating between many
different narrows in different ways.