eslint: Disable spellcheck on generated JS code.
The preceding commit generated some lint warnings -- oops.  The
generated code uses the `instanceof` operator... and because the whole
thing is a giant ES6 template literal consisting of JS/ES3 code, the
spell-checker applies normal-English spelling rules to it.

Just turn that rule off on that file; it's not source code, so we
really don't care about spelling in it.