android: Fix label of intent filter.
This `android:label` attribute is "used when [the app or activity] is
presented to the user as having the capability described by the
filter".  In other words it's a user-facing message -- so the current
value is not at all a good one.

Concretely, in this case, it's used in the OAuth functionality.  So
after selecting account in the browser, user is redirected to the app
with this intent. If multiple apps are capable of taking this intent
(like in our case when both release and debug builds are installed in
the phone), then a pop up is shown by the phone OS, letting user to
select with which app he wants to proceed. And this label is used in
this pop up. `filter_react_native` is totally irrelevant, so change it
to something relevant, which can be used to differentiate both the
apps in the phone.

Fortunately, this message never seems to actually be shown unless two
different things advertise this "capability", which in this case
basically only happens when both the release and debug builds of the
app are installed.  So only us developers of the app actually see it.
Still, no need to have a confusing value like this one.

The default behavior is simply to take `android:label` from the parent
component (here, the activity), or else from the application.  Those
are set nicely to "Zulip" or "Zulip (debug)"; so just leave it at that.

[greg: revised explanation in commit message]