ComposeBox: Allow showing a thumbnail preview of images before sending.
This commit makes the necessary changes to ComposeMenu and ComposeBox
in order to select and upload images per message while composing while
showing a thumbnail preview. It also allows you to delete a selected
image before sending the message and choose another one.

The code in this commit is written to handle multiple images but the
max limit is currently set to 1. Can enable more in the future.

It also allows you to select a topic for an image.