webView, js: Return endNode id more reliably from `getStartAndEndNodes`.
In this method, a try is made to get `endNode` at cordinate (200,
window.innerHeight - 20), but if message list is short then at this
position there is not any node. So in that case when node is undefined
then take help of `getNearestMsgNodeUsing` to get `endNode`. Call it
with `lastChild` of the `documentBody` and propogate using
`previousSibling`.

This also fixes the issue of marking a unseen message as read, i.e if
at that cordinate (bottom) if there is time row then before msg below
the timerow was marked as read. Now if there is time row, or typing or
any other node then from that propogate up (using `previousSibling`)
to get nearest message node.

Propogate using `endNode` if it is timeRow or some div (i.e not body
or document, on which propogation is not possible) else use the
lastChild of the documentBody.

With this extra care, only seen messages are marked as read.

Fixes: #2988