ComposeMenu: Extract default filename for new image picker.
This commit adds a function to extract the filename from an image
uri. This will be useful in the following commit when we switch
to a new image picker library. On android, the filename is currently
not returned by the react-native-image-crop-picker so we edit
chooseUploadImageFilename to be able to infer the filename
from the uri instead.