animatedScaleComponent: Remove use of `display` property.
It works same as it works in CSS. There is no use of such
property in animation component, as they result in a sudden change.

It was used to make size 0 whenever component is not supposed to show.
But if there is nothing to show then it's size will be automatically
0.

This will not cause any issue. It has been used in unread notice and
autocomplete popup. Refer previous commit, which is a better
alternative for this in unread notice.

For autocomplete, when it is not suppossed to pop, it's children will
be null and nothing will be rendered. Thus taking no space (refer
0baca8f). Part of #2943.