SpinningProgress: Replace temporarily with SpinningProgressImg
React Native's ART library is important and we probably will use
it more in the future. It is the only good option when we want to
achieve some visual elements / styles that we would use CSS or
SVG on the web instead.

Unfortunately, there is a serious bug on Android, that have existed
for a while. Surprisingly we did not encounter this bug until RN 0.57
but it has always been there.

Visually, the issue manifests as black spinning square instead of
the usual progress spinner.

Bug reported to React Native here:
https://github.com/facebook/react-native/issues/17565

A possible solution looks something like:
https://github.com/macdoum1/react-native/commit/587848168a80a4cf2244f7c7f6aaff397c675c58

Possible alternatives evaluated:
 * Styles (ala CSS) - previous implementation used that, but was not
reliable between versions (due to RN issues)
 * use third-party component - a bit too complicated, and they suffer
from the same issues. For example:
https://github.com/bartgryszko/react-native-circular-progress/pull/105
 * use static images - that is the chosen workaround

The exact look was recreated in Sketch and exported as PNG. We use
two versions - one with 'brand' color and one 'black'. We are losing
the functionality to change 'thickness' which we don't use.

Once the bug in ART is fixed upstream, we will revert this change.